backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.0.4
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
joblib==1.1.0
networkx==2.7
nltk==3.7
numpy==1.22.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
regex==2022.3.2
scipy==1.8.0
sqlparse==0.4.2
tqdm==4.63.0
whitenoise==6.0.0
